pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-translate"
version = "0.1.0"
description = "PDF academic paper split, chunk, and translate with local memory"
requires-python = ">=3.10"
dependencies = [
  "pymupdf>=1.24.0",
  "typer>=0.12.0",
  "pyyaml>=6.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "bcrypt>=4.0.0",
  "PyJWT>=2.8.0",
]

[project.scripts]
pdf-translate = "pdf_translate.cli:main"
pdf-translate-web = "pdf_translate.server.__main__:main"

[tool.setuptools]
packages = [
  "pdf_translate",
  "pdf_translate.translators",
  "pdf_translate.server",
]
